Добавил проверки Update для hole и part
В сервисах добавил методы update и добавил проверку что есть что изменять.
В API методах update заменил save на update
Привел все id к int

+++ src/main/java/dto/HoleDto.java
@@ -12,9 +12,9 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class HoleDto {
 
-    Integer holeId;
+    int holeId;
 
-    Integer partId;
+    long partId;
 
     double diameter;
